модернизация проекта Miles

+++ src/Main.java
@@ -1,8 +1,7 @@
 public class Main {
     public static void main(String[] args) {
-        float mile = 20;
-        float price = 8456;
-        float bonus = price / mile;
+        int mile = 20;
+        int price = 8456;
+        int bonus = price / mile;
         System.out.println(bonus);
-    }
-}
+    }}
